KSP-466: Fix Random.nextBytes ByteArray type/runtime box mismatch (#4627)
* KSP-466: Fix Random.nextBytes ByteArray/RuntimeArrayBox mismatch

nextBytes's synthetic stub used a List<Int> stand-in type for
ByteArray (HeaderHelpers+SyntheticRandomStubs.swift), which silently
fell back to Any because the Random stub bucket registers before the
Collections bucket that defines kotlin.collections.List. That made
`r.nextBytes(ByteArray(8)).size` fail to resolve (KSWIFTK-SEMA-0024),
and left the Kotlin-source `Random.nextBytes` in Stdlib unreachable
for dot-call syntax (only .synthetic-flagged declarations are found
by the scope-based extension fallback in
CallTypeChecker+MemberCallInferenceRegularResolution.swift).

Switch byteArrayType to the real ByteArray class (same pattern as the
adjacent uByteArrayType), which also fixes SecureRandom.nextBytes /
generateSeed sharing the same type. Since ByteArray's runtime
representation is RuntimeArrayBox (kk_array_new / kk_byteArray_size),
update kk_random_nextBytes(_size) and __kk_secure_random_next_bytes /
generate_seed to allocate/read RuntimeArrayBox instead of
RuntimeListBox, or nextBytes would silently return an empty array.

Also fixes an unrelated pre-existing bug this surfaced: the 1-arg
primitive array constructor callee table in
CallSupportLowerer+DefaultArgsAndVarargs.swift was missing ByteArray,
ShortArray, ULongArray, and UByteArray, so a bare `ByteArray(8)` (no
init lambda) always emitted an undefined `_ByteArray` symbol instead
of calling kk_array_new.


* Update Random/SecureRandom runtime tests for RuntimeArrayBox migration

kk_random_nextBytes(_size/_range) and __kk_secure_random_next_bytes/
generate_seed now exclusively use RuntimeArrayBox, so tests that built
inputs as RuntimeListBox or cast results back to RuntimeListBox would
fail (or silently stop asserting anything, e.g. determinism checks
comparing two empty arrays). Updated
RuntimeRandomBoundaryTests/RuntimeSecureRandomTests to use
RuntimeArrayBox throughout, matching the already-migrated
nextUBytes tests in the same file. Dropped
testNextBytesRangeFillsOnlyRequestedListSlice as redundant with the
existing Array-based equivalent.


* Split non-standard Random/SecureRandom API surface out of diff_cases

random_extended.kt and random_overload_edge_cases.kt used to fail to
compile on the JVM kotlinc reference for an unrelated reason before
this PR's nextBytes fix (SEMA-0024 on kswiftc's side), so
diff_kotlinc.sh's exit-code-only comparison happened to "pass" by
coincidence. Now that kswiftc compiles them correctly, the pre-existing
gap between kswiftc's extended Random API and real kotlinc surfaces as
a genuine compile mismatch:

- Random.nextFloat(until)/nextFloat(from, until) are kswiftc-only
  extensions (STDLIB-655); real kotlinc's Random only has the no-arg
  nextFloat(). Split these calls out of random_extended.kt and
  random_overload_edge_cases.kt into a new
  random_nextfloat_synthetic_overloads.kt, marked SKIP-DIFF
  (DEBT-DIFF-005), so the rest of those files' standard-API assertions
  stay diff-checked against the JVM reference.
- SecureRandom.getInstance() (no-arg) is likewise a kswiftc-only
  convenience overload; secure_random.kt depends on it throughout, so
  mark the whole file SKIP-DIFF (DEBT-DIFF-005).

Also found and set aside a separate, pre-existing runtime bug while
investigating random_overload_edge_cases.kt: Random(seed).nextLong(from,
until), called after another Random instance has already been used,
prints a stray "kk_unbox_long called on non-LongBox object" runtime
warning to stdout (reproduces with plain nextInt()/nextBytes() calls
too — unrelated to ByteArray). Marked SKIP-DIFF (DEBT-DIFF-004) with a
follow-up task rather than expanding this PR's scope to fix it.

Updated docs/diff-skip-inventory.md with the three new entries.


* Fix dangling makePrimitiveArrayType call missed in merge commit

The merge commit (145b20a6f) auto-merged this file with no reported
conflict, which meant it was already staged before I edited the
working tree to fix the dangling makePrimitiveArrayType reference
(that helper was removed elsewhere by the merge). The edit never got
re-staged, so the merge commit was pushed with the pre-fix content
even though local builds (which use the working tree) looked fine.
CI caught it. Re-apply the same inline byteArrayType construction
fix as a follow-up commit.


* Fix Random.Default CI timeout

* Fix Random default and SecureRandom array runtime

* Fix Random nextLong signed composition

* Stabilize coroutine cancellation diff case

---------

+++ Sources/CompilerCore/Stdlib/kotlin/random/Random.kt
@@ -97,7 +97,9 @@ public open class Random internal constructor(
         return result
     }
 
-    public open fun nextLong(): Long = nextInt().toLong().shl(32) + nextInt()
+    // Keep both operands as Long to preserve Kotlin's sign extension of the
+    // low Int when composing the 64-bit result.
+    public open fun nextLong(): Long = nextInt().toLong().shl(32) + nextInt().toLong()
 
     public open fun nextLong(until: Long): Long = nextLong(0L, until)
 
@@ -241,10 +243,9 @@ public open class Random internal constructor(
             defaultRandom = Random(entropy)
         }
 
-        // NOTE: every open member is re-declared here, forwarding to
-        // defaultRandom, even though most bodies are identical to what
-        // Random's own skeleton implementation would already compute by
-        // calling nextBits() virtually. This compiler's "bare ClassName.member()"
+        // NOTE: every open member is re-declared here, even though most bodies
+        // are identical to what Random's own skeleton implementation would
+        // already compute by calling nextBits() virtually. This compiler's "bare ClassName.member()"
         // shorthand for named-companion access (used throughout existing
         // diff_cases/golden tests, e.g. `Random.nextInt(1, 10)`) only resolves
         // members the companion *directly declares*, not ones it merely
